actionablescience-agentic-studio-ed7a12c24fe1/backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
fastapi==0.118.2
h11==0.16.0
httpx==0.27.0
idna==3.10
langgraph==0.2.59
langchain-core==0.3.29
langchain-openai==0.2.14
openai>=1.58.1
psycopg2-binary==2.9.10
pydantic==2.12.0
pydantic_core==2.41.1
python-dotenv==1.1.1
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
